+++ earth2mip/networks/graphcast/channels.py
@@ -2,12 +2,8 @@
 from typing import List
 from earth2mip.initial_conditions import cds
 import numpy as np
-import dataclasses
 from modulus.utils.sfno.zenith_angle import cos_zenith_angle
 from graphcast.graphcast import TaskConfig
-import pandas as pd
-import graphcast.data_utils
-import pandas
 import xarray
 
 
@@ -99,103 +95,6 @@ time_dependent = {
 }
 
 
-channels = []
-code_to_channel = dict(zip(cds.CHANNEL_TO_CODE.values(), cds.CHANNEL_TO_CODE.keys()))
-
-
-def yield_channels():
-    for v in pl_inputs:
-        code = pl_inputs[v]
-        for level in levels:
-            yield code_to_channel[code] + str(level)
-
-    for v in sl_inputs:
-        code = sl_inputs[v]
-        yield code_to_channel[code]
-
-
-def yield_channels_ecmwf():
-    for v in pl_inputs:
-        code = pl_inputs[v]
-        for level in levels:
-            yield code_to_channel[code], code, level
-
-    for v in sl_inputs:
-        code = sl_inputs[v]
-        yield code_to_channel[code], code, None
-
-
-def unpack(x, codes, pressure_level_codes, single_level_codes, levels):
-    """retrieve variable from x
-
-    Args:
-        x: array (b, h, c, y, x)
-        channels: names of c dim
-        levels: levels to get
-        pressure_level_codes_to_get
-        single_level_codes_to_get
-    """
-    idx = pandas.Index(codes)
-    output = xarray.Dataset()
-    for code in pressure_level_codes:
-        name = CODE_TO_GRAPHCAST_NAME[code]
-        indexer = idx.get_indexer(
-            [cds.PressureLevelCode(code, level) for level in levels]
-        )
-        output[name] = ["batch", "time", "level", "lat", "lon"], x[:, :, indexer]
-    for code in single_level_codes:
-        name = CODE_TO_GRAPHCAST_NAME[code]
-        indexer = codes.index(cds.SingleLevelCode(code))
-        output[name] = ["batch", "time", "lat", "lon"], x[:, :, indexer]
-    return output.assign_coords(level=levels)
-
-
-def unpack_all(x, codes):
-
-    p_codes = set()
-    s_codes = set()
-    levels = set()
-
-    for c in codes:
-        if isinstance(c, cds.SingleLevelCode):
-            s_codes.add(c.id)
-        elif isinstance(c, cds.PressureLevelCode):
-            p_codes.add(c.id)
-            levels.add(c.level)
-
-    return unpack(x, codes, sorted(p_codes), sorted(s_codes), levels=sorted(levels))
-
-
-def assign_grid_coords(ds, grid):
-    coords = {
-        "lat": grid.lat[::-1],
-        "lon": grid.lon,
-    }
-    return ds.assign_coords(coords)
-
-
-def pack(ds: xarray.Dataset, codes) -> np.ndarray:
-    """retrieve variable from x
-
-    Args:
-        x: array (b, h, c, y, x)
-        channels: names of c dim
-        levels: levels to get
-        pressure_level_codes_to_get
-        single_level_codes_to_get
-    """
-    idx = pandas.Index(codes)
-    shape = (1, ds.sizes["time"], len(codes), ds.sizes["lat"], ds.sizes["lon"])
-    x = np.zeros(shape=shape)
-    for k, code in enumerate(codes):
-        name = CODE_TO_GRAPHCAST_NAME[code.id]
-        if isinstance(code, cds.SingleLevelCode):
-            x[:, :, k] = ds[name]
-        elif isinstance(code, cds.PressureLevelCode):
-            x[:, :, k] = ds[name].sel(level=code.level)
-    return x
-
-
 def is_3d(name):
     return name in pl_inputs
 
@@ -230,27 +129,6 @@ def toa_incident_solar_radiation(time, lat, lon):
     solar_constant = 1361  #  W/m²
     z = cos_zenith_angle(time, lon, lat)
     return np.maximum(0, z) * solar_constant
-
-
-def add_dynamic_vars(ds, time):
-    if "batch" in ds.dims:
-        assert ds.sizes["batch"] == 1, "haven't vectorized cos zenith yet over time"
-    ds.coords["init_time"] = xarray.Variable(["batch"], [pd.Timestamp(time)])
-    ds.coords["datetime"] = ds.init_time + ds.time
-    graphcast.data_utils.add_derived_vars(ds)
-    del ds["year_progress"]
-    del ds["day_progress"]
-
-    toas = []
-    for dt in ds.time.values.flat:
-        tt = pd.Timestamp(time) - pd.Timedelta(dt)
-        datetime = tt.to_pydatetime()
-        toas.append(toa_incident_solar_radiation(time, ds.lat, ds.lon))
-    ds["toa_incident_solar_radiation"] = xarray.concat(toas, dim="time").expand_dims(
-        "batch"
-    )
-    del ds.coords["datetime"]
-    del ds.coords["init_time"]
 
 
 def get_data_for_code_scalar(code, scalar):
